Refactor: tipdos e interfaces en su archivo formulario seccion asociados
